refactor: refactored modals and handlers completely out of ImportPage.jsx
